Improved test coverage around regex hash
This commit also contains some style cleanups, a new gem,
minitest/focus. And a change to RegexHash to raise a no Method
error when a string cannot be matched.  I attempted to raise KeyError,
but that bubbled up to no method error and didn't behave as I expected.
Todo: read more about the exception call stack